template setup change + only show best templates
