+++ src/app/projects/[id]/generate/workspace/GenerateSidebar.tsx
@@ -58,7 +58,7 @@ interface PanelOption {
 const navItems: WorkspacePanelG[] = [
   { type: 'chat', id: 'chat', name: "Chat", icon: MessageSquareIcon, href: "#chat" },
   { type: 'preview', id: 'preview', name: "Preview", icon: PlayIcon, href: "#preview" },
-  // { type: 'templates', id: 'templates', name: "Templates", icon: LayoutTemplateIcon, href: "#templates" },
+  { type: 'templates', id: 'templates', name: "Templates", icon: LayoutTemplateIcon, href: "#templates" },
   { type: 'code', id: 'code', name: "Code", icon: Code2Icon, href: "#code" },
 ];
 
@@ -78,6 +78,14 @@ const PANEL_OPTIONS: PanelOption[] = [
     color: 'from-green-500 to-green-600',
   },
   {
+    type: 'templates',
+    label: 'Templates',
+    description: 'Professional pre-made templates',
+    icon: <LayoutTemplateIcon className="h-5 w-5" />,
+    badge: 'NEW',
+    color: 'from-pink-500 to-pink-600',
+  },
+  {
     type: 'storyboard',
     label: 'Storyboard',
     description: 'Visual timeline of your scenes',
@@ -91,16 +99,6 @@ const PANEL_OPTIONS: PanelOption[] = [
     icon: <Code2Icon className="h-5 w-5" />,
     color: 'from-gray-500 to-gray-600',
   },
-  /*
-  {
-    type: 'templates',
-    label: 'Templates',
-    description: 'Professional pre-made templates',
-    icon: <LayoutTemplateIcon className="h-5 w-5" />,
-    badge: 'NEW',
-    color: 'from-pink-500 to-pink-600',
-  },
-  */
 ];
 
 export function GenerateSidebar({
