+++ src/app/projects/[id]/generate/workspace/WorkspaceContentAreaG.tsx
@@ -179,6 +179,17 @@ function DropZoneG({
       )
     },
     { 
+      type: 'templates', 
+      label: 'Templates', 
+      icon: (
+        <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+          <rect width="18" height="7" x="3" y="3" rx="1"/>
+          <rect width="9" height="7" x="3" y="14" rx="1"/>
+          <rect width="5" height="7" x="16" y="14" rx="1"/>
+        </svg>
+      )
+    },
+    { 
       type: 'storyboard', 
       label: 'Storyboard', 
       icon: (
